# schemas/openapi.yaml
openapi: "3.1.0"
info:
  title: cxt Sync API
  version: "1"
  description: |
    REST API of the central server for cxt (cxthub).
    This file is the **common wire contract** for the cli, backend, and frontend components.
    In a fully separated architecture (cli / backend do not share Go modules), this file
    serves as the unique shared contract role.
    Full description: docs/SYNC-PROTOCOL.md. Domain entities: docs/DATA-MODEL.md.
    CIR object schema: schemas/cir.schema.json. Manifest schema: schemas/manifest.schema.json.
  contact:
    name: cxthub team
    email: alice@example.com
servers:
  - url: "https://{host}/api/v1"
    description: Central server (main/dev common)
    variables:
      host:
        default: localhost:8080
        description: Server host. Local development is localhost:8080, production is team domain.

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
security:
  - SessionCookie: []
  - BearerToken: []

# Permission model (5-step role ladder — serial AND gates):
#   Authentication → workspace membership → role (viewer→puller→member→maintainer→owner) → action-specific policies.
#   viewer     read-only (web context·graph)          — GET repos/snapshots/docs/memories/diff
#   puller     +local synchronization of team assets  — GET settings/secrets/settings-objects, pull/objects
#   member     +Context write (default)               — push/*, PUT refs/memories/settings-objects, fork
#   maintainer +Team asset write·invite creation          — PUT settings/secrets, PATCH about, POST invites
#   owner      +Operations                              — PATCH workspace, member management, transfer(creator)
# viewer level GET is allowed for anonymous if public workspace. Policy(secrets/settings_policy) is
# maintainer gate can narrow down to 'owner only'.
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: cxt_session
      description: |
        Web login session(HttpOnly cookie). POST /auth/session exchanges IDP token(Firebase/dev) for
        issuance. Token value is not exposed in JS.
    BearerToken:
      type: http
      scheme: bearer
      description: |
        Authorization: Bearer <token>. Accepts two formats:
        - `sess_*` server session token — CLI token(sess_cli_*, POST /me/cli-tokens issuance, 1 year) included.
          CLI stores `cxt login <token>` in ~/.cxt/auth.json(0600, per host).
        - IDP token(Firebase ID token / dev token) — validated for user upsert.
...
  schemas:
...
    ErrorResponse:
      type: object
      description: "Common error envelope. docs/SYNC-PROTOCOL.md §8."
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              enum:
                - bad_request
                - unauthenticated
                - forbidden
                - team_mismatch
                - not_found
                - non_fast_forward
                - ref_conflict
                - git_origin_mismatch
                - conflict
                - validation
                - integrity_violation
                - missing_object
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
  responses:
    Unauthorized:
      description: "No token or invalid (unauthenticated)."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /auth/device/approve:
    post:
      operationId: deviceApprove
      summary: pairing approval — web login user confirms code
      tags: [identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
      responses:
        "200": { description: approved }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { description: "expired/nonexistent code" }
  /auth/device/poll:
    post:
      operationId: devicePoll
      summary: Approval polling — Receive CLI token once upon approval
      security: []
      tags: [identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, poll_token]
              properties:
                code: { type: string }
                poll_token: { type: string }
      responses:
        "200":
          description: "pending or approved (+token — Exposed only once in this response)"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [pending, approved] }
                  token: { type: string }
                  expires_at: { type: string, format: date-time }
        "404": { description: "Expired/Invalid pairing (including permission mismatch — Information not exposed)" }
...